move bitmask_loader to to_coco.py
